refactor(goja): Extract gojaClass to separate file
